+++ app/models/transform/Transform.java
@@ -2,6 +2,6 @@ package models.transform;
 
 //------------------------------------------------------------
 public interface Transform {
-	public Double apply(Double result);
+	public Float apply(Float result);
 	public String debug();
 }
